+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ### v1.7
 - Migrate bigdecimal-**: constants (TWO), getNumberOfDigitsAfterTheDecimalPointOrZero() method
 - Fix bug with @Mandatory + "get" in the middle of the property: target1 -> getTarget1() is expected getter, not tar1()
+- Migrate collection utility methods: get mutable list, set and map
 
 ### v1.6
 - Add "countryCode" -> GeoLocation